Jenkinsfile parameterize build

+++ Jenkinsfile
@@ -1,15 +1,11 @@
 pipeline {
     agent none
     parameters {
-        string(name: 'BC_LIB', defaultValue: '', description: 'braincloud-unreal-plugin-src branch (blank for .gitmodules)')
+        //string(name: 'BC_LIB', defaultValue: '', description: 'braincloud-unreal-plugin-src branch (blank for .gitmodules)')
         string(name: 'BRANCH_NAME', defaultValue: 'develop', description: 'examples-unreal branch')
         booleanParam(name: 'CLEAN_BUILD', defaultValue: true, description: 'clean pull and build')
-        booleanParam(name: 'RelayTestApp', defaultValue: true, description: 'build?')
-        booleanParam(name: 'RelayTestAppCPP', defaultValue: false, description: 'build?')
-        booleanParam(name: 'TappyChicken', defaultValue: false, description: 'build?')
-        booleanParam(name: 'ScriptTestApp', defaultValue: false, description: 'build?')
-        booleanParam(name: 'Groups', defaultValue: false, description: 'build?')
-        booleanParam(name: 'Leaderboard', defaultValue: false, description: 'build?')
+        choice(name: 'PROJECTNAME', choices: ['all', 'RelayTestApp', 'RelayTestAppCPP', 'TappyChicken', 'ScriptTestApp','Groups', 'Leaderboard'], description: 'Where to run tests?')
+        //choice(name: 'PLATFORM', choices: ['all', 'Mac', 'Win64', 'IOS', 'Android'])
     }
     stages {
 
@@ -29,29 +25,22 @@ pipeline {
                 if (${params.CLEAN_BUILD}) {
                     deleteDir()
                 }
-                checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])				
-                sh 'autobuild/checkout-submodule.sh ${BC_LIB}'
+                checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
+                //sh 'autobuild/checkout-submodule.sh ${BC_LIB}' //not ported to windows yet
                 sh 'autobuild/_brainCloudSetup_examples-unreal.command'
-                if(${params.RelayTestApp}) {
+                if(${params.PROJECTNAME == 'all'}) {
                     sh 'autobuild/makebuild.sh RelayTestApp MAC UE_5_Mac'
+                    sh 'autobuild/makebuild.sh RelayTestAppCpp MAC UE_5_Mac'
+                    sh 'autobuild/makebuild.sh TappyChicken MAC UE_5_Mac'
+                    sh 'autobuild/makebuild.sh ScriptTestApp MAC UE_5_Mac'
+                    sh 'autobuild/makebuild.sh Groups MAC UE_5_Mac'
+                    sh 'autobuild/makebuild.sh Leaderboard MAC UE_5_Mac'
                     //sh 'autobuild/makebuild.sh RelayTestApp IOS'
                     //sh 'autobuild/makebuild.sh RelayTestApp ANDROID'
                 }
-                if(${params.RelayTestAppCPP}) {
-                    sh 'autobuild/makebuild.sh RelayTestAppCpp MAC UE_5_Mac'
+                else{
+                    sh 'autobuild/makebuild.sh ${params.PROJECTNAME} MAC UE_5_Mac'
                 }
-                if(${params.TappyChicken}) {
-                    sh 'autobuild/makebuild.sh TappyChicken MAC UE_5_Mac'
-                }
-                if(${params.ScriptTestApp}) {
-                    sh 'autobuild/makebuild.sh ScriptTestApp MAC UE_5_Mac'
-                if(${params.Groups}) {
-                    sh 'autobuild/makebuild.sh Groups MAC UE_5_Mac'
-                }
-                if(${params.Leaderboard}) {
-                    sh 'autobuild/makebuild.sh Leaderboard MAC UE_5_Mac'
-                }
-
             }
             post {
                 success {
@@ -72,22 +61,27 @@ pipeline {
               }
              steps {
                  echo "---- braincloud Code Pull ${BRANCH_NAME} ${BC_LIB}"
-                 //if (${CLEAN_BUILD}) {
+                 if (${params.CLEAN_BUILD}) {
                     deleteDir()
-                 //}
+                 }
                  checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
                  bat 'autobuild\\_brainCloudSetup_examples-unreal.bat'
-                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_5_Win'
-                 //bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win'
-                 //b/at 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat Groups Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat Leaderboard Win64 UE_4_Win'
+                if(${params.PROJECTNAME == 'all'}) {
+                    bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_5_Win64'
+                    bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_5_Win64'
+                    bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_5_Win64'
+                    bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_5_Win64'
+                    bat 'autobuild\\makebuild.bat Groups Win64 UE_5_Win64'
+                    bat 'autobuild\\makebuild.bat Leaderboard Win64 UE_5_Win64'
+                }
+                else{
+                    bat 'autobuild\\makebuild.bat ${params.PROJECTNAME} Win64 UE_4_Win64'
+                }
             }
             post {
                 success {
-                    fileOperations([fileZipOperation(folderPath: 'UE_5_Win', outputFolderPath: '.')])
-                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_5_Win.zip', followSymlinks: false, onlyIfSuccessful: true
+                    fileOperations([fileZipOperation(folderPath: 'UE_5_Win64', outputFolderPath: '.')])
+                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_5_Win64.zip', followSymlinks: false, onlyIfSuccessful: true
                 }
             }
        }
@@ -103,20 +97,25 @@ pipeline {
               }
              steps {
                  echo "---- braincloud Code Pull ue4-examples ${BC_LIB}"
-                 //if (${CLEAN_BUILD}) {
+                 if (${params.CLEAN_BUILD}) {
                      deleteDir()
-                 //}
+                 }
                  checkout([$class: 'GitSCM', branches: [[name: '*/ue4-examples']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
                  bat 'autobuild\\_brainCloudSetup_examples-unreal-4.bat'
-                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_4_Win'
-                // bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win'
+                 if(${params.PROJECTNAME == 'all'}) {
+                    bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_4_Win64'
+                    bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Wi64'
+                    bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win64'
+                    bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win64'
+                 }
+                else{
+                    bat 'autobuild\\makebuild.bat ${params.PROJECTNAME} Win64 UE_4_Win64'
+                }
              }
             post {
                 success {
-                    fileOperations([fileZipOperation(folderPath: 'UE_4_Win', outputFolderPath: '.')])
-                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_4_Win.zip', followSymlinks: false, onlyIfSuccessful: true
+                    fileOperations([fileZipOperation(folderPath: 'UE_4_Win64', outputFolderPath: '.')])
+                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_4_Win64.zip', followSymlinks: false, onlyIfSuccessful: true
                 }
             }
        }
